extracted layout to separate component, proper icon nav
