+++ src/App.tsx
@@ -1,10 +1,8 @@
 import React from "react";
-import { Link, Outlet, Route, Routes } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import "twin.macro";
 import DemoThing from "./DemoThing";
-import MetricsIcon from "./icons/MetricsIcon";
-import ProfilesIcon from "./icons/ProfilesIcon";
-import SettingsIcon from "./icons/SettingsIcon";
+import Layout from "./Layout";
 import Metrics from "./Metrics";
 
 const App = () => {
@@ -32,34 +30,6 @@ const Settings = () => (
 const Profiles = () => (
   <div>
     <h1>All profiles</h1>
-  </div>
-);
-
-const Layout = () => (
-  <div tw="h-screen font-lab-grotesque light:(bg-off-white text-darker-grey) dark:(bg-dark-grey text-lighter-grey)">
-    <nav tw="w-full fixed bottom-0 light:(bg-off-white bg-opacity-80 backdrop-blur text-darker-grey) dark:(bg-dark-grey bg-opacity-80 backdrop-blur text-lighter-grey)">
-      <ul tw="flex flex-row pt-8 pb-8">
-        <li tw="flex-1 text-center">
-          <Link to="/settings">
-            <SettingsIcon />
-          </Link>
-        </li>
-        <li tw="flex-1 text-center">
-          <Link to="/">
-            <MetricsIcon />
-          </Link>
-        </li>
-        <li tw="flex-1 text-center">
-          <Link to="/profiles">
-            <ProfilesIcon />
-          </Link>
-        </li>
-      </ul>
-    </nav>
-
-    <main tw="p-14">
-      <Outlet />
-    </main>
   </div>
 );
