+++ src/icons/ProfilesIcon.tsx
@@ -4,7 +4,6 @@ import "twin.macro";
 const ProfilesIcon = () => (
   <svg
     tw="stroke-current"
-    viewBox="0 0 24 24"
     width="24"
     height="24"
     fill="none"
